refactor(utils): add getDiffInMilliseconds as ESM
