+++ src/utils/index.ts
@@ -1,0 +1,1 @@
+export { default as getDiffInMilliseconds } from './getDiffInMilliseconds.js';
